+++ Editor/PropertyDrawers/MonoBehaviourEditor.cs
@@ -1,0 +1,9 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace LurkingNinja.Utils.Editor.PropertyDrawers
+{
+    [CanEditMultipleObjects,
+     CustomEditor(typeof(MonoBehaviour), true)]
+    public class MonoBehaviourEditor : UnityEditor.Editor {}
+}