+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2022-12-03
+### Added
+- Optional type
+- Optional property drawer
+- Value type to store different value types
+- Bits field to store individual bits compressed into byte-arrays
+- IntRange
+- IntRange property drawer
+- FloatRange
+- FloatRange property drawer
+- ValueRange
+- ScriptableObject property drawer to enable drawing SOs as properties on MonoBehaviours
+
 ## [1.0.0] - 2022-12-03
 ### Added
 - CHANGELOG (this file)